Chunks are downloaded if they are not on the device!

+++ app/src/main/java/com/aaronnebbs/peersplitandroidapplication/Model/ChunkFile.java
@@ -5,15 +5,15 @@ import java.io.File;
 public class ChunkFile {
     private File file;
     private String originalname;
-    private int size;
+    private long size;
 
-    public ChunkFile(File file, String name, int size){
+    public ChunkFile(File file, String name, long size){
         this.file = file;
         this.originalname = name;
         this.size = size;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
